Restore game

+++ YSI.CurseOfSilverCrown.EndOfTurn/EndOfTurnTask.cs
@@ -2,21 +2,21 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
-using YSI.CurseOfSilverCrown.EndOfTurn.Actions;
-using YSI.CurseOfSilverCrown.Core.Database.EF;
-using YSI.CurseOfSilverCrown.Core.Database.Models;
-using YSI.CurseOfSilverCrown.Core.Database.Enums;
-using YSI.CurseOfSilverCrown.Core.Interfaces;
 using YSI.CurseOfSilverCrown.Core;
+using YSI.CurseOfSilverCrown.Core.Database.EF;
+using YSI.CurseOfSilverCrown.Core.Database.Enums;
+using YSI.CurseOfSilverCrown.Core.Database.Models;
+using YSI.CurseOfSilverCrown.Core.Database.Models.GameWorld;
 using YSI.CurseOfSilverCrown.Core.Helpers;
+using YSI.CurseOfSilverCrown.Core.Interfaces;
 using YSI.CurseOfSilverCrown.Core.Parameters;
+using YSI.CurseOfSilverCrown.EndOfTurn.Actions;
 
 namespace YSI.CurseOfSilverCrown.EndOfTurn
 {
     public class EndOfTurnTask
     {
-        private ApplicationDbContext Context;
+        private readonly ApplicationDbContext Context;
         private Turn CurrentTurn;
 
         private int eventNumber;
@@ -46,8 +46,19 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
             }
         }
 
+        private void UpdateEventNumber()
+        {
+            eventNumber = 0;
+            var eventsOfCurrentTurn = Context.EventStories
+                .Where(e => e.TurnId == CurrentTurn.Id);
+            if (eventsOfCurrentTurn.Any())
+                eventNumber = eventsOfCurrentTurn.Max(e => e.Id) + 1;
+        }
+
         private void RunCommands()
         {
+            UpdateEventNumber();
+
             var runCommands = Context.Commands.ToList();
             var organizations = Context.Domains.ToArray();
 
@@ -81,8 +92,8 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
             {
                 var unit = Context.Units.Find(unitId);
                 if (unit.Type == enArmyCommandType.WarSupportDefense &&
-                    unit.TargetDomainId == unit.PositionDomainId || 
-                    unit.Type == enArmyCommandType.CollectTax && 
+                    unit.TargetDomainId == unit.PositionDomainId ||
+                    unit.Type == enArmyCommandType.CollectTax &&
                     unit.DomainId == unit.PositionDomainId)
                 {
                     unit.Status = enCommandStatus.Complited;
@@ -100,7 +111,9 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
                         .Single(u => u.Id == unitId);
                     if (unit.Status == enCommandStatus.Complited ||
                         unit.ActionPoints < WarConstants.ActionPointsFullCount - subTurn * WarConstants.ActionPointForMoveWarriors)
+                    {
                         continue;
+                    }
 
                     switch (unit.Type)
                     {
@@ -191,7 +204,7 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
                 var unit = Context.Units.Find(unitId);
                 unit.Status = enCommandStatus.Complited;
                 Context.Update(unit);
-            }            
+            }
 
             var unitForDelete = Context.Units.Where(c => c.Warriors <= 0 || c.Status == enCommandStatus.Destroyed);
             Context.RemoveRange(unitForDelete);
@@ -217,7 +230,9 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
                 var unitPosition = Context.Domains.Find(unit.PositionDomainId.Value);
                 if (KingdomHelper.IsSameKingdoms(Context.Domains, unitDomain, unitPosition) ||
                     DomainRelationsHelper.HasPermissionOfPassage(Context, unitDomain.Id, unitPosition.Id))
+                {
                     continue;
+                }
 
                 unit.Status = enCommandStatus.Retreat;
                 Context.Update(unit);
@@ -287,37 +302,55 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
             CreatorCommandForNewTurn.CreateNewCommandsForOrganizations(Context, domains);
         }
 
-        private void PrepareCommands()
+        private List<Domain> GetDomainsForPrepareCommands()
         {
-            var domains = Context.Domains
+            return Context.Domains
                 .Include(o => o.Person)
                 .Include("Person.User")
                 .Include(o => o.Suzerain)
                 .Include(o => o.Commands)
                 .ToList();
+        }
+
+        private int? GetInitiatorRunIdForPrepareCommands(IEnumerable<IGrouping<int, Command>> groups, Domain domain)
+        {
+            if (!groups.Any())
+            {
+                return null;
+            }
+            else if (groups.Count() > 1)
+            {
+                var domainIsActive = domain.Person.User != null &&
+                                     domain.Person.User.LastActivityTime > DateTime.UtcNow - new TimeSpan(24, 0, 0);
+                return domainIsActive
+                    ? domain.PersonId
+                    : domain.Suzerain.PersonId;
+            }
+            else
+            {
+                return domain.PersonId;
+            }
+        }
+
+        private void PrepareCommands()
+        {
+            var domains = GetDomainsForPrepareCommands();
             foreach (var domain in domains)
             {
-                var initiatorRunId = 0;
+
                 var groups = domain.Commands
                     .GroupBy(c => c.InitiatorPersonId);
-                if (groups.Count() > 1)
-                {
-                    var domainIsActive = domain.Person.User != null &&
-                                         domain.Person.User.LastActivityTime > DateTime.UtcNow - new TimeSpan(24, 0, 0);
-                    initiatorRunId = domainIsActive
-                        ? domain.PersonId
-                        : domain.Suzerain.PersonId;
-                }
-                else
-                    initiatorRunId = domain.PersonId;
+                var initiatorRunId = GetInitiatorRunIdForPrepareCommands(groups, domain);
+                if (initiatorRunId == null)
+                    continue;
 
                 var groupsForDelete = groups.Where(g => g.Key != initiatorRunId);
                 foreach (var group in groupsForDelete)
                     Context.RemoveRange(group.ToList());
 
                 var groupForRun = groups.Single(g => g.Key == initiatorRunId);
-                foreach (var command in groupForRun)                
-                    command.InitiatorPersonId = domain.PersonId;                
+                foreach (var command in groupForRun)
+                    command.InitiatorPersonId = domain.PersonId;
                 Context.UpdateRange(groupForRun);
                 Context.SaveChanges();
             }
@@ -345,7 +378,9 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
                         : domain.Suzerain.PersonId;
                 }
                 else
+                {
                     initiatorRunId = domain.PersonId;
+                }
 
                 var groupsForDelete = groups.Where(g => g.Key != initiatorRunId);
                 foreach (var group in groupsForDelete)
@@ -371,9 +406,11 @@ namespace YSI.CurseOfSilverCrown.EndOfTurn
                 Context.SaveChanges();
             }
             else
+            {
                 CurrentTurn = Context.Turns
                     .OrderByDescending(t => t.Id)
                     .First();
+            }
         }
 
         private void ExecuteGoldTransferAction(Turn currentTurn, IEnumerable<Command> currentCommands)
